src/backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bandit==1.8.3
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
dateparser==1.2.1
Deprecated==1.2.18
dnspython==2.7.0
einops==0.8.1
email_validator==2.2.0
fastapi==0.115.7
fastapi-cli==0.0.7
filelock==3.17.0
fonttools==4.57.0
fsspec==2025.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1rc1
google-api-python-client==2.159.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
limits==4.7.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.3
packaging==24.2
pbr==6.1.1
pillow==11.1.0
proto-plus==1.26.0rc1
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pymongo==4.11.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
setuptools==75.8.2
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.45.3
stevedore==5.4.1
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.48.3
typer==0.15.1
typing_extensions==4.12.2
tzlocal==5.3
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.2
wheel==0.45.1
wrapt==1.17.2
